Add QR code scanning functionality
